requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
lightfm>=1.17
torch>=2.0.0
transformers>=4.21.0
librosa>=0.10.0
soundfile>=0.12.1
faiss-cpu>=1.7.4
prometheus-client>=0.17.0
psutil>=5.9.0
pandas>=2.0.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests>=2.31.0
httpx>=0.28.0
pathlib2>=2.3.0
typing-extensions>=4.7.0
click>=8.1.0
